messaging_app/requirements.txt
# Django and related packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database connectors
mysqlclient==2.2.0
psycopg2-binary==2.9.9

# Authentication and security
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
cryptography==41.0.7

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Utilities
requests==2.31.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1

# Development tools
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0
